+++ releases.md
@@ -1,5 +1,5 @@
  * 5.13
-   * Add last recent dirs when open worksheets
+   * Add last recent dir when open worksheets
    * Fix when xlsx file has no item detected
  * 5.12
    * Fix hide of tooltip when not focus.